+++ tests/FunctionAllTest.php
@@ -68,30 +68,4 @@ class FunctionAllTest extends TestCase
         all([resolve(1), reject(2), resolve(3)])
             ->then($this->expectCallableNever(), $mock);
     }
-
-    /** @test */
-    public function shouldAcceptAPromiseForAnArray()
-    {
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->once())
-            ->method('__invoke')
-            ->with($this->identicalTo([1, 2, 3]));
-
-        all(resolve([1, 2, 3]))
-            ->then($mock);
-    }
-
-    /** @test */
-    public function shouldResolveToEmptyArrayWhenInputPromiseDoesNotResolveToArray()
-    {
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->once())
-            ->method('__invoke')
-            ->with($this->identicalTo([]));
-
-        all(resolve(1))
-            ->then($mock);
-    }
 }