+++ tests/FunctionMapTest.php
@@ -79,36 +79,6 @@ class FunctionMapTest extends TestCase
     }
 
     /** @test */
-    public function shouldAcceptAPromiseForAnArray()
-    {
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->once())
-            ->method('__invoke')
-            ->with($this->identicalTo([2, 4, 6]));
-
-        map(
-            resolve([1, resolve(2), 3]),
-            $this->mapper()
-        )->then($mock);
-    }
-
-    /** @test */
-    public function shouldResolveToEmptyArrayWhenInputPromiseDoesNotResolveToArray()
-    {
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->once())
-            ->method('__invoke')
-            ->with($this->identicalTo([]));
-
-        map(
-            resolve(1),
-            $this->mapper()
-        )->then($mock);
-    }
-
-    /** @test */
     public function shouldRejectWhenInputContainsRejection()
     {
         $mock = $this->createCallableMock();
@@ -121,35 +91,6 @@ class FunctionMapTest extends TestCase
             [resolve(1), reject(2), resolve(3)],
             $this->mapper()
         )->then($this->expectCallableNever(), $mock);
-    }
-
-    /** @test */
-    public function shouldRejectWhenInputPromiseRejects()
-    {
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->once())
-            ->method('__invoke')
-            ->with($this->identicalTo(null));
-
-        map(
-            reject(),
-            $this->mapper()
-        )->then($this->expectCallableNever(), $mock);
-    }
-
-    /** @test */
-    public function shouldCancelInputPromise()
-    {
-        $mock = $this->getMock('React\Promise\CancellablePromiseInterface');
-        $mock
-            ->expects($this->once())
-            ->method('cancel');
-
-        map(
-            $mock,
-            $this->mapper()
-        )->cancel();
     }
 
     /** @test */